Fix apex-ctl clap usage and update it
Update clap to 4.0.26. Enable the `derive` feature to fix compilation.

Change the derive attributes to work with newer version of clap.

Drop apex_hardware::FareBuffer use.

+++ apex-ctl/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
-use apex_hardware::{Device, FrameBuffer, USBDevice};
-use clap::Parser;
+use apex_hardware::{Device, USBDevice};
+use clap::{ArgAction, Parser, Subcommand};
 use log::{info, LevelFilter};
 use simplelog::{Config as LoggerConfig, SimpleLogger};
 
@@ -8,13 +8,13 @@ use simplelog::{Config as LoggerConfig, SimpleLogger};
 #[clap(version = "1.0", author = "not-jan")]
 struct Opts {
     /// A level of verbosity, and can be used multiple times
-    #[clap(short, long, parse(from_occurrences))]
-    verbose: i32,
-    #[clap(subcommand)]
+    #[arg(short, long, action = ArgAction::Count)]
+    verbose: u8,
+    #[command(subcommand)]
     subcmd: SubCommand,
 }
 
-#[derive(Parser)]
+#[derive(Subcommand)]
 enum SubCommand {
     /// Clear the OLED screen
     Clear,
